lint debt cleared: every one of the ~200 tracked findings is paid and the gate goes fully uncapped (run.max-issues-per-linter: 0, max-same-issues: 0) so nothing hides behind the default 3-identical-issues rounds again - golangci-lint run ./... answers 0 issues; the real fixes: the eh-eh tautology in the east seam scan, the dead end clamp in chordWalksRegion, the always-true autonomous param and the always-nil error of runSessionSupervised (six call sites simplified to send nil after the supervisor answers), five dead declarations deleted (soakMaxLevel, zoneOverrideSlack, Cell.patrolZone, Mesh.dstComps, runHopBudget), five always-constant test helper params folded into the helpers (huntTile, buildRealRegion, assembleTile, navmeshPost, loadRegionForDump), the G109 bounds guard moved next to the conversion, errorlint now checks errors.Is on the io.ErrUnexpectedEOF pairs, the union-find var+assign pairs merged (S1021), the concat loops rebuilt on strings.Builder and strings.Join (perfsprint), the wasted assignment and the ineffectual assignment dropped, ClosestPoint and the zone index carry the honest blank params, the v1/v2 wire decode twins and the attack request packet carry documented dupl relief, the C1 water zone table keeps its one-zone-per-line layout under an lll exclusion with the reason, the two replay probes join the analysis-driver relief, cyclop joins the test exclusions with the scenario-script reason; the suppression policy is uniform: 36 complexity sites carry one //nolint:a,b // note directly above the func with the refactor note inside the 80 columns, the deliberate partial inits live in the exhaustruct_v5 ignore-patterns list grouped by reason (empty answers, option aggregates, internal build state), nolintlint audited out every stale directive (nine removed, two moved onto the line the finding actually reports); AGENTS.md gains the Tree cleanliness discipline section (caps off, task fmt before commit, the suppression policy, relief lives in the config, dead code is deleted) and the Code conventions section matches the real config (gci disabled, the complexity and exhaustruct policies as enforced); tools/prepush.sh and the CI lint step flip from lint --new to the full uncapped lint and the workflow lands at .github/workflows/ci.yml (the push needs the workflow-scoped token, docs/ci_workflow.yml stays the owner copy)

+++ internal/swarm/pathfind/navmesh/astar.go
@@ -77,7 +77,7 @@ func newZoneIndex(boxes []WaterZone) *zoneIndex {
         y1 := int64(box.MaxY) >> zoneCellShift
         for x := x0; x <= x1; x++ {
             for y := y0; y <= y1; y++ {
-                key := uint64(uint64(x)<<32) | uint64(uint32(y))
+                key := uint64(x)<<32 | uint64(uint32(y))
                 idx.cells[key] = append(idx.cells[key], box)
             }
         }
@@ -93,7 +93,7 @@ func newZoneIndex(boxes []WaterZone) *zoneIndex {
 // swimming character floats near the surface inside the box, the
 // zone data minZ never splits a water body the surface covers).
 func (idx *zoneIndex) covered(x, y, z float64) bool {
-    key := uint64(uint64(int64(x)>>zoneCellShift)<<32) |
+    key := uint64(int64(x)>>zoneCellShift)<<32 |
         uint64(uint32(int64(y)>>zoneCellShift))
     for _, box := range idx.cells[key] {
         if x >= box.MinX && x <= box.MaxX && y >= box.MinY &&
@@ -538,6 +538,8 @@ func (m *Mesh) astar(
 // context walls the links onto banned polygons and prices the escape
 // polygons of the ban holding the start (the recovery ban rules of
 // the grid costTo - the rectangle granularity form, see avoid.go).
+//
+//nolint:cyclop // the astar expansion dispatches per neighbor kind
 func (m *Mesh) expand(state *queryState, node *astarNode, idx uint32,
     endPos Pos, filter Filter, avoid avoidCtx, allow *confinedSet,
 ) {
